Giving up on porting this file over
I got pretty close to pulling this python implementation straight into
C#, but without broadcasting it gets pretty difficult. I'm not unwilling
to add a broadcasting function, but it feels like a lot of work for an
implementation that I don't really like that much. I've learned a lot
about python, C#, importing libraries like DLL's into Unity, and linear
algebra/neural nets, so I think this was a win.